fix(dx): split 3 files over the linecheck 500-line gate (#974)
main's CI is currently red: the linecheck job (.github/workflows/lint.yml) fails
because src/cli/mod.rs (535 lines), src/cli/tests.rs (526 lines), and
src/routines/ical_tests.rs (590 lines) have all grown past the 500-line-per-file
gate. Splits each along an existing seam — bind-address/loopback policy into
src/cli/bind.rs (+ bind_tests.rs), and the svc_ical/svc_ical_routine/build_ical
service-layer tests into src/routines/ical_service_tests.rs — mirroring the
existing #[path] submodule pattern already used throughout cli/ and ical_tests.rs.
No behavior change; fmt/clippy/tests/coverage/linecheck all pass locally.
